Cross-platform validation: macOS green end to end, AU gated, three new guards
WHAT RESTORING macOS ACTUALLY REVEALED. The compile break fixed in 6f63573 was
three days of a red job. What it COST was three rounds of KI-009 work: the
broken line sits INSIDE testClipMixCannotChangeTheDefaultPresetsSound, so every
regression written in rounds 5, 6 and 7 — written for a fault that reproduces
only on macOS — had never once executed on macOS. "The suites are green" meant
"green on the platforms that still ran", and that is the platform the report is
not about. POSTMORTEMS.md INC-003 records both halves; the second is the
expensive one.

With the build restored the whole macOS job is green (run 31435190370,
macos-14, universal arm64 + x86_64): build -> both suites -> pluginval x3 in
BOTH modes at the build.yml strictness -> packaging. So the first arm64
execution of every previously-dark test passes, and the answer is a negative
one: the suites do not reproduce KI-009 on macOS either. That is a stronger
negative than round 7's — measured on the reproducing platform rather than
inferred from the other one — and it still means the reproduction lives outside
what the suites drive.

THE ONE REAL DEFECT CLASS THE AUDIT FOUND, pinned rather than described. At
Clip Mix 0 the mix loop's exact endpoints leave the dry sample untouched, so
ClipSat's OUTPUT is finite whatever its internals did — while tameLp[ch] is
updated on BOTH tame branches to keep the 6 kHz split warm. Warm state plus an
invisible output is a latch, because invariant 9's repair is keyed on
stageGeneratedNonFinite and the BOUNDARY is what raises it. That shape is the
reported fingerprint exactly: channel-local, gated on a non-zero mix, absent at
mix 0, upstream of the limiter's detector tap, cured by bypass. It is NOT
reachable in the shipped build — 0.1.3's colour-argument bound holds the stage —
and testClipSatCannotHideANonFiniteFromTheBoundary now proves that over 30
poisoning attempts up to FLT_MAX with the colour stage swept. Remove the bound
and the same test fails with 32 000 of 120 000 non-finite samples ON ORDINARY
AUDIO, after the poisoning stopped: a deferred, state-latched form of the same
fault that no existing test could catch, because at mix 0 there is nothing at
the boundary to catch.

THE REGRESSION THE REPORT ASKED FOR. A pushed chain x every non-zero Clip Mix,
which nothing ran: round 5 swept the mix at the DEFAULT preset, where clipDrive
exact-skips the clipper and every branch provably lands on the same float, and
the loudness cases pushed the chain at one mix value. 1 % is in the sweep
because a barely-open mix is the HARDEST case for a mix-scaled fault, not the
easiest. Each case now asserts its own premise (clipDrive actually engaged) so
it cannot go vacuous the way round 6 found the whole battery had. Mutation-
verified with a partial-mix-only channel loss — a mutant the pre-existing rows
cannot see, since the default mix is 100 %.

HYPOTHESES KILLED BY EVIDENCE, not argument:
  * uninitialised/stale per-channel state — both suites are valgrind-clean under
    memcheck with --track-origins (0 errors from 0 contexts) and clean under
    ASan+UBSan. The defect would be platform-independent even though its symptom
    is not, so the tools see it whatever bytes the allocator supplied.
  * wrapper, bus layout, channel pointers — bypass feeds ONLY the engine's
    output selector; no DSP is skipped, no state reset. "Bypass restores the
    left channel" therefore PROVES the host delivers channel 0 and the dry ring
    is intact.
  * the two macOS -Wshorten-64-to-32 warnings in the region access — AudioBlock
    getSample/setSample take int; value-preserving round trip, already removed.

CI, expanded to answer three different questions rather than one:
  * source-lint — scripts/check-portability.py. A LINT because no Linux build
    can see the class: GCC and Clang both accept the line that breaks macOS.
  * linux-clang — the AppleClang diagnostic set GCC does not apply, gated on
    first-party paths (not -Werror: JUCE's own sources compile into our targets,
    and a blanket -Werror gates on a dependency and gets switched off). Also
    runs the compile canary that checks the pinned JUCE still HAS the hazard.
    It does NOT catch the typedef class, and the docs say so.
  * sanitizers — ASan+UBSan over both suites, valgrind memcheck over the DSP
    suite. MemorySanitizer is deliberately not used: it needs every dependency
    instrumented, and an uninstrumented one yields false positives, not silence.
  * macOS now validates the AU as well as the VST3 — same strictness, both
    modes, 3 passes. The AU is the format KI-009 is reported in and the one that
    had passed no automated validation at all. `au` on a non-Darwin host is an
    error, not a silent skip.

Four first-party Clang warnings cleared so the new gate is green on arrival
(two int64_t->double, two int->float, and the editor's `processor` member, which
shadowed AudioProcessorEditor::processor and is now `proc`). All bit-identical.

Suites 295 + 743 = 1038, green on Linux under GCC, under Clang, under ASan+UBSan
and under valgrind — and green on macOS arm64.

+++ src/dsp/AdaptiveEngine.h
@@ -488,8 +488,8 @@ public:
         // deliberately no signal back to the wrapper at P4.
         if (learnBlocks > 0)
         {
-            const float ons  = (float) (learnOnsSum  / learnBlocks);
-            const float tilt = (float) (learnTiltSum / learnBlocks);
+            const float ons  = (float) (learnOnsSum  / (double) learnBlocks);
+            const float tilt = (float) (learnTiltSum / (double) learnBlocks);
             // A pass that measured through an overflow is REFUSED here rather
             // than repaired later, and the check lives at the writer for a
             // reason: sanitiseState() cancels such a pass, but it runs later in
